Centralize user management tests in a single file
